# ♻️Refactor: reduce DefaultCtx from 736 to 728 bytes