--- a/ctx_interface.go
+++ b/ctx_interface.go
@@ -19,7 +19,7 @@ type CustomCtx interface {
 	// Methods to use with next stack.
 	getMethodINT() int
 	getIndexRoute() int
-	getTreePath() string
+	getTreePathHash() int
 	getDetectionPath() string
 	getPathOriginal() string
 	getValues() *[maxParams]string